requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
djangorestframework==3.12.2
mysqlclient==2.0.2
Pillow==8.0.1
python-decouple==3.3
xlsxwriter==1.3.7   
django_crontab==0.7.1
python-dateutil==2.8.1
numpy==1.19.5
pandas==1.1.5
xhtml2pdf==0.2.5
